streamlit/requirements.txt
pip
Cython
pyproj
numpy
pandas
geopandas
plotly-express
Shapely
osmnx
streamlit